ndg::config

Module templates

Enums

TemplateError

Error type for template operations.

Constants

DEFAULT_JSON_TEMPLATE

Default configuration template in JSON format.

DEFAULT_TOML_TEMPLATE

Default configuration template in TOML. We’ve also included sufficient amount of comments to explain each field so that the user is not immediately lost. Yuck though, I do not like having to embed a string one bit.

Functions

get_template

Get the correct configuration template based on the requested format.