ndg

Module manpage

Modules

options 🔒

Statics

TROFF_ESCAPE

TROFF_FORMATTING

Functions

escape_leading_dot

Escape a leading dot to prevent it from being interpreted as a troff command

generate_manpage

Generate a manpage from options JSON Generate a manpage from options JSON

get_roff_escapes

Map of characters that need to be escaped in manpages

man_escape

Escapes a string for use in manpages