ndg::utils

Module markdown

Functions

collect_included_files

Collects all included files from markdown documents in the input directory.

create_processor_from_config

Creates a markdown processor from the NDG configuration.

extract_page_title

Extracts the page title from a markdown file.

process_markdown_files

Processes all markdown files in the input directory and writes HTML output.